+++ src/sdk/index.ts
@@ -1,3 +1,4 @@
 export * from './getArchiveFolders';
 export * from './getArchives';
 export * from './getFolder';
+export * from './getRecord';